requirements.txt
# =============================================================================
# OpenDoc - Dependencies
# =============================================================================

# --- LLM Engine (Gemini) ---
google-generativeai>=0.8.0
google-auth>=2.34.0

# --- Browser Automation (SAT portal, insurance portals) ---
playwright>=1.48.0

# --- Database ---
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.35

# --- Web Framework (Dashboard API) ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# --- Data Processing ---
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# --- PDF Generation (Prescriptions, reports) ---
reportlab>=4.2.0
Pillow>=10.4.0

# --- HTTP Client ---
httpx>=0.27.0

# --- Cryptography (e.firma certificates, XML signing for SAT SOAP) ---
cryptography>=42.0.0
lxml>=5.1.0

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.24.0
